fixed logger
