+++ test/www/jxcore/lib/testUtils.js
@@ -10,7 +10,6 @@ var path = require('path');
 var randomString = require('randomstring');
 var Promise = require('bluebird');
 var https = require('https');
-var Logger = require('thali/thaliLogger');
 var logger = require('thali/thaliLogger')('testUtils');
 var ForeverAgent = require('forever-agent');
 var thaliConfig = require('thali/NextGeneration/thaliConfig');
@@ -80,42 +79,6 @@ function isFunction(functionToCheck) {
   return functionToCheck && getType.toString.call(functionToCheck) ===
     '[object Function]';
 }
-
-(function () {
-  var logCallback;
-  var messages = [];
-
-  if (isMobile) {
-    Mobile('setLogCallback').registerAsync(function (callback) {
-      logCallback = callback;
-      messages.forEach(function (message) {
-        logCallback(message);
-      });
-      messages = [];
-    });
-  }
-
-  /**
-  * Log a message to the screen - only applies when running on Mobile.
-  * It assumes we are using our test framework with our Cordova WebView
-  * who is setup to receive logging messages and display them.
-  */
-  module.exports.getLogger = function (meta) {
-    var logger = Logger(meta);
-    if (isMobile) {
-      logger._thaliLogger.on('message', function (message) {
-        // 'logCallback' is required.
-        // We should save all messages before it will be available.
-        if (logCallback) {
-          logCallback(message);
-        } else {
-          messages.push(message);
-        }
-      });
-    }
-    return logger;
-  }
-}) ();
 
 var myName = '';
 var myNameCallback = null;
